Improve /knock capabilities and restrict its usage in lobby areas
/knock now sends blank IC message with knock ding to players in the
same area as the player who initiates the knock, as well as their
intended destination. Deaf people only get the blank IC message.
For that, crucially, the act of actually sending MS packets was
move to a client function as opposed to being handled by aoprotocol

/knock also now does not work if the user is in an area that has the
tag 'lobby_area: true', or attempts to knock into an area that was
similarly defined.

Affected tests were also rewritten to comply with this.