Make /scream show messages in IC+Refactor IC send method
The actual send IC method is now a client method that can be called
in places other than aoprotocol. This allows for greater flexibility
when trying to send custom IC messages to other clients, such as the
ones for narration (which are used for /knock and /narrate) or the
ones that involve an "unknown sender" (which is used for /scream)

Also made the tester explicitly keep track of whether it has at
some point discarded an IC message for its mock client as a result
of said client receiving the same message/character combo twice,
which allows for easier debugging.