Make /help be able to give specific command help+Improve README.md
Using /help with a command name will now give instructions on how
to use said command as per README.md according to the rank of the
user. It will try and return the highest rank description for the
user. If the user is not allowed to use the command themselves,
they will get the least-rank description for the command, as well
as the rank they need to use it.

README.md was also revamped to remove tab characters, as well as
improving certain command descriptions and reshuffling some of them
(mostly the login/logout commands) so that they align better with
the minimum rank they require.