Add /zone+Improve /ban /kick notifications+Added `` to some messages
/zone was filled out to work at its barest minimum. Test cases were
created to assert functionality, and the zone module was fitted
with a method to assert structural integrity during debugging.

Ban, banhdid and kick were improved to notify all present mods and
cms of the action being performed (previously it was just to the
people in the area of the infractor).

Some messages will include backticks to separate their names from
a sentence ending period where such a period could be reasonably
be interpreted to be part of a name.