Add /zone_remove+/zone_delete + Make new zone exceptions + Other test stuff
/zone_remove will remove an area from the user's watched zone by ID
or name.

/zone_delete will delete the user's watched zone entirely.

ZoneManager now has two new exceptions that are triggered if either an
area or watcher are assumed part of a zone's area or watcher set when
in reality they are not.

Introduced type hints for some fields in structures.py to help with
IDEs autocompleting (why did no one tell me about this before?)