Add /zone_play+Make /play&/rplay loop+Add new server no music exception
/zone_plÃay will send an order to all clients in the zone a user is
watching to play a specified track, regardless of whether the track
exists or not in the server music list.

/play, /rplay and /zone_play will now loop tracks if the given track
is part of the server music list.

Added ServerError.NoMusicFoundError to deal with the specific case
where the user tries to play custom music.

Moved the handling of playing a track in an area into a wrapper
function in AreaManager.Area (play_track), which eliminates the need
to have the same duplicated methods in aoprotocol, and some of the
command methods such as /play and /rplay.