Restrict elevated RP notifications to staff zone watchers+Other stuff
All RP notifications will now be limited in scope to staff members
who are also watching the zone the event took place at, even if they
themselves are not in an area part of the zone the event took place at.

Allowed /zone_unwatch to be used by non-staff in case of /logout
prior to /zone_unwatch.