Deal with code that would raise SyntaxWarning in P3.8+Establish support for P3.8
The /roll and /rollp code would raise SyntaxWarning on console in case of
bad user input of the form /roll 20 3(5-1), where the user expected the
modifier to do a multiplication between 3 and (5-1). Python does not like
this sort of multiplication and was correctly caught as TypeError before.
However, Python 3.8 would raise a console warning too as part of its
TypeError catching, so now the code asserts that the modifier does not
have this form before attempting to evaluate it.

Moreover, Python updating instructions were added to the Readme, and
Python 3.8 support was explicitly established in there.