Private servers will now include the masterserver name for IP distribution
When launching a server, TsuserverDR provides a handy IP+Port+Name
combination to distribute to players so that they can connect to the
server without need of using the master server list. However, servers
that did not wish to be listed in the master server did not include
the name in this handy link, instead keeping it as blank. Now, it will
include such name.

Also updated the changelog to reflect the fact Python 3.8 is now
explicitly supported.