Send notifications to zwatchers if character change for any reason+Fix rogue messages
Notifications are now sent to zone watchers if a player in an area part of a zone
changes character, let it be by their own choice or due to conflict due to existing/
restricted characters.

Fixed showname update notifications being sent when the change took place in an area
not part of a zone.