Send notifications to zwatchers if player in zone/zwatcher dc+Send zone delete messages
If a player in a zone disconnects, or a watcher of a zone disconnects, the remaining
zone watchers will get a notification warning them about this.

Moreover, if a zone has no more watchers, it will now send an announcement to all mods
and CMs about the automatic zone deletion.