Notify zwatchers on sneak/reveal+Revamp notification system for zwatchers
Zone watchers will now receive a message if someone does /sneak or /reveal on a
player. Likewise, they will receive a message telling them if the player revealed
themselves by speaking IC or playing music.

Moreover, the entire notification system for zone watchers was rebuilt so that
proper distinction could be made in the cases where it matters whether to send
a notification to staff members if the sender is not in an area part of a zone.
Now there is a rebuilt is_zstaff and is_zstaff_flex parameters for build_cond
to handle STRICT requirement of being in a zone to receive the message and a
"flexible" requirement of being in a zone respectively.