Rework client shutdown for Python 3.8+Fix /help help message+Fix shutdown grammar issue
The server shutdown process was reworked for Python 3.8. Previously, it would kick out
the clients out of the areas and allow for the asyncio processes to be garbage collected
and discarded. However, Python 3.8 raises an "ignored exception" issue on server shutdown
when the server is closed like this. Now, the server manually disconnects all clients
before shutting itself down.

/help help now indicates that /help points to the server repository, not "this" (which
would make sense if reading from the README.md file, not as part of the command).

The server now properly indicated "Kicking 1 remaining client" as opposed to "Kicking 1
remaining clients" (note the "s") when shutting down a server with only 1 client connected.