Implement variety of bug fixes, mostly related to server operation and zones
* Fixed /zone_watch and /zone_delete not catching the KeyError that ZoneManager throws
when passing in a zone ID that does not exist. The exception is now rethrown as a
ZoneError, which is correctly caught by the server and parsed accordingly.

* Fixed rare issue with AWS instances attempting to ping api.ipify.org to obtain their
public IP addresses by providing an SSL context when attempting to select the URL.

* If the server unrecoverably crashes due to a Python error or an incorrect setup,
TsuserverDR will now expect the operator to Press Enter to exit. This is to account
for server operators who launch their servers via clicking start_server.py, which
without this, had their windows immediately closed, so they could not see what
the issue was.

* Error log files are now created if the server crashes on bootup. A slightly
modified log file is produced that does not include area and client dump, because,
well, the server did not bootup. In any other aspect, they are identical to the
error log files produced by other server errors.

* Made sure a master server connection, if any, was closed when the server shuts down.
This is to prevent this shutting down taking place as part of Python's garbage collection,
which is an issue for Python 3.8 as by the time the garbage collector comes in, the asyncio
loop is already closed and in this version of Python, this raises a RuntimeError. As the
exception is thrown as part of the garbage collection, it is not a crash per-se, but a
server operator should not be expected to see such exception appear when they shut down
their server.

This is also a test of the GitHub issue tracker.