Add config manager
Specifying the config fields in code lets IDEs and tools see
which config options are available, and using config will be
less prone to bugs. The rationale is that requesting config
should never ever break anything and the application should
not have to check if the config exists for every time it uses it.
There should always exist some sane default we can give the application.

A test should be made in order to keep the config manager and default config file
in sync.