Make IP property of client again and log it on connection
While it's perfectly understandable to not show IPs to mods
or players, it's bog standard practice to log the IP address
on connections. Every webserver does this.

Since client.ip was actually a getter for ipid, replace
every instance of client.ip with ipid.

Will note that ipid is a very misleading name. It's a numeric, unique
ID of each client and in principle it has nothing to do with IP.